move form data functions to seperate service
