+++ src/app/calculator/system-and-equipment-curve/curve-data.service.spec.ts
@@ -1,0 +1,12 @@
+import { TestBed } from '@angular/core/testing';
+
+import { CurveDataService } from './curve-data.service';
+
+describe('CurveDataService', () => {
+  beforeEach(() => TestBed.configureTestingModule({}));
+
+  it('should be created', () => {
+    const service: CurveDataService = TestBed.get(CurveDataService);
+    expect(service).toBeTruthy();
+  });
+});
