+++ src/app/calculator/system-and-equipment-curve/equipment-curve/equipment-curve-form/equipment-curve-form.component.ts
@@ -4,6 +4,7 @@ import { FormGroup } from '@angular/forms';
 import { EquipmentCurveService } from '../equipment-curve.service';
 import { SystemAndEquipmentCurveService, EquipmentInputs } from '../../system-and-equipment-curve.service';
 import { Subscription } from 'rxjs';
+import { CurveDataService } from '../../curve-data.service';
 
 @Component({
   selector: 'app-equipment-curve-form',
@@ -25,12 +26,13 @@ export class EquipmentCurveFormComponent implements OnInit {
   ];
   modWarning: string = null;
   resetFormsSub: Subscription;
-  constructor(private equipmentCurveService: EquipmentCurveService, private systemAndEquipmentCurveService: SystemAndEquipmentCurveService) { }
+  constructor(private equipmentCurveService: EquipmentCurveService, private systemAndEquipmentCurveService: SystemAndEquipmentCurveService,
+    private curveDataService: CurveDataService) { }
 
   ngOnInit() {
     this.initForm();
     this.setSmallUnit();
-    this.resetFormsSub = this.systemAndEquipmentCurveService.resetForms.subscribe(val => {
+    this.resetFormsSub = this.curveDataService.resetForms.subscribe(val => {
       if (val == true) {
         this.resetForm();
       }
