+++ src/app/calculator/system-and-equipment-curve/system-curve/fan-system-curve-form/fan-system-curve-form.component.ts
@@ -4,6 +4,7 @@ import { FanSystemCurveFormService } from '../fan-system-curve-form.service';
 import { Settings } from '../../../../shared/models/settings';
 import { SystemAndEquipmentCurveService, FanSystemCurveData } from '../../system-and-equipment-curve.service';
 import { Subscription } from 'rxjs';
+import { CurveDataService } from '../../curve-data.service';
 
 @Component({
   selector: 'app-fan-system-curve-form',
@@ -21,11 +22,12 @@ export class FanSystemCurveFormComponent implements OnInit {
   resetFormsSub: Subscription;
   assessmentDataPoints: Array<{ pointName: string, flowRate: number, yValue: number }>;
   showDataPointOptions: boolean = false;
-  constructor(private fanSystemCurveFormService: FanSystemCurveFormService, private systemAndEquipmentCurveService: SystemAndEquipmentCurveService) { }
+  constructor(private fanSystemCurveFormService: FanSystemCurveFormService, private systemAndEquipmentCurveService: SystemAndEquipmentCurveService,
+    private curveDataService: CurveDataService) { }
 
   ngOnInit() {
     this.initForm();
-    this.resetFormsSub = this.systemAndEquipmentCurveService.resetForms.subscribe(val => {
+    this.resetFormsSub = this.curveDataService.resetForms.subscribe(val => {
       if (val == true) {
         this.initForm();
       }
