+++ src/app/calculator/system-and-equipment-curve/system-curve/pump-system-curve-form/pump-system-curve-form.component.ts
@@ -4,6 +4,7 @@ import { PumpSystemCurveFormService } from '../pump-system-curve-form.service';
 import { Settings } from '../../../../shared/models/settings';
 import { SystemAndEquipmentCurveService, PumpSystemCurveData } from '../../system-and-equipment-curve.service';
 import { Subscription } from 'rxjs';
+import { CurveDataService } from '../../curve-data.service';
 
 @Component({
   selector: 'app-pump-system-curve-form',
@@ -20,11 +21,12 @@ export class PumpSystemCurveFormComponent implements OnInit {
   resetFormsSub: Subscription;
   assessmentDataPoints: Array<{ pointName: string, flowRate: number, yValue: number }>;
   showDataPointOptions: boolean = false;
-  constructor(private pumpSystemCurveFormService: PumpSystemCurveFormService, private systemAndEquipmentCurveService: SystemAndEquipmentCurveService) { }
+  constructor(private pumpSystemCurveFormService: PumpSystemCurveFormService, private systemAndEquipmentCurveService: SystemAndEquipmentCurveService,
+    private curveDataService: CurveDataService) { }
 
   ngOnInit() {
     this.initForm();
-    this.resetFormsSub = this.systemAndEquipmentCurveService.resetForms.subscribe(val => {
+    this.resetFormsSub = this.curveDataService.resetForms.subscribe(val => {
       if (val == true) {
         this.initForm();
       }
